move findTargetRule to base config
